+++ cucumber-expressions/javascript/src/GeneratedExpression.ts
@@ -8,10 +8,7 @@ export default class GeneratedExpression {
   ) {}
 
   get source() {
-    return util.format(
-      this.expressionTemplate,
-      ...this.parameterTypes.map((t) => t.name)
-    )
+    return util.format(this.expressionTemplate, ...this.parameterTypes.map((t) => t.name))
   }
 
   /**
@@ -21,16 +18,11 @@ export default class GeneratedExpression {
    */
   get parameterNames(): ReadonlyArray<string> {
     const usageByTypeName: { [key: string]: number } = {}
-    return this.parameterTypes.map((t) =>
-      getParameterName(t.name, usageByTypeName)
-    )
+    return this.parameterTypes.map((t) => getParameterName(t.name, usageByTypeName))
   }
 }
 
-function getParameterName(
-  typeName: string,
-  usageByTypeName: { [key: string]: number }
-) {
+function getParameterName(typeName: string, usageByTypeName: { [key: string]: number }) {
   let count = usageByTypeName[typeName]
   count = count ? count + 1 : 1
   usageByTypeName[typeName] = count