+++ fake-cucumber/javascript/src/ErrorMessageGenerator.ts
@@ -1,9 +1,6 @@
 import StackUtils from 'stack-utils'
 
-export type MakeErrorMessage = (
-  error: Error,
-  sourceFrames: ReadonlyArray<string>
-) => string
+export type MakeErrorMessage = (error: Error, sourceFrames: ReadonlyArray<string>) => string
 
 export function withFullStackTrace(): MakeErrorMessage {
   const stack = new StackUtils({