+++ gherkin-utils/javascript/test/QueryTest.ts
@@ -1,4 +1,4 @@
-import GherkinStreams from '@cucumber/gherkin/dist/src/stream/GherkinStreams'
+import { GherkinStreams } from '@cucumber/gherkin-streams'
 import { IdGenerator, messages } from '@cucumber/messages'
 import { pipeline, Readable, Writable } from 'stream'
 import assert from 'assert'
@@ -55,12 +55,8 @@ describe('Query', () => {
 `
       )
 
-      const gherkinDocument = envelopes.find(
-        (envelope) => envelope.gherkinDocument
-      ).gherkinDocument
-      const scenario = gherkinDocument.feature.children.find(
-        (child) => child.scenario
-      ).scenario
+      const gherkinDocument = envelopes.find((envelope) => envelope.gherkinDocument).gherkinDocument
+      const scenario = gherkinDocument.feature.children.find((child) => child.scenario).scenario
 
       const pickleId = envelopes.find((e) => e.pickle).pickle.id
       const pickleIds = gherkinQuery.getPickleIds('test.feature', scenario.id)
@@ -77,9 +73,7 @@ describe('Query', () => {
     Given a passed step
 `
       )
-      const expectedPickleIds = envelopes
-        .filter((e) => e.pickle)
-        .map((e) => e.pickle.id)
+      const expectedPickleIds = envelopes.filter((e) => e.pickle).map((e) => e.pickle.id)
       const pickleIds = gherkinQuery.getPickleIds('test.feature')
       assert.deepStrictEqual(pickleIds, expectedPickleIds)
     })
@@ -95,12 +89,9 @@ describe('Query', () => {
 `
       )
 
-      assert.throws(
-        () => gherkinQuery.getPickleIds('test.feature', 'some-non-existing-id'),
-        {
-          message: 'No values found for key 6. Keys: [some-non-existing-id]',
-        }
-      )
+      assert.throws(() => gherkinQuery.getPickleIds('test.feature', 'some-non-existing-id'), {
+        message: 'No values found for key 6. Keys: [some-non-existing-id]',
+      })
     })
 
     it('avoids dupes and ignores empty scenarios', async () => {
@@ -148,8 +139,8 @@ describe('Query', () => {
         .find((envelope) => envelope.pickle)
         .pickle.steps.map((pickleStep) => pickleStep.id)
 
-      const stepId = envelopes.find((envelope) => envelope.gherkinDocument)
-        .gherkinDocument.feature.children[0].scenario.steps[0].id
+      const stepId = envelopes.find((envelope) => envelope.gherkinDocument).gherkinDocument.feature
+        .children[0].scenario.steps[0].id
 
       assert.deepEqual(gherkinQuery.getPickleStepIds(stepId), pickleStepIds)
     })
@@ -169,18 +160,14 @@ describe('Query', () => {
   `
         )
 
-        const backgroundStepId = envelopes.find(
-          (envelope) => envelope.gherkinDocument
-        ).gherkinDocument.feature.children[0].background.steps[0].id
+        const backgroundStepId = envelopes.find((envelope) => envelope.gherkinDocument)
+          .gherkinDocument.feature.children[0].background.steps[0].id
 
         const pickleStepIds = envelopes
           .filter((envelope) => envelope.pickle)
           .map((envelope) => envelope.pickle.steps[0].id)
 
-        assert.deepEqual(
-          gherkinQuery.getPickleStepIds(backgroundStepId),
-          pickleStepIds
-        )
+        assert.deepEqual(gherkinQuery.getPickleStepIds(backgroundStepId), pickleStepIds)
       })
 
       it('return all pickleStepIds linked to a step in a scenario with examples', async () => {
@@ -197,18 +184,14 @@ describe('Query', () => {
 `
         )
 
-        const scenarioStepId = envelopes.find(
-          (envelope) => envelope.gherkinDocument
-        ).gherkinDocument.feature.children[0].scenario.steps[1].id
+        const scenarioStepId = envelopes.find((envelope) => envelope.gherkinDocument)
+          .gherkinDocument.feature.children[0].scenario.steps[1].id
 
         const pickleStepIds = envelopes
           .filter((envelope) => envelope.pickle)
           .map((envelope) => envelope.pickle.steps[1].id)
 
-        assert.deepEqual(
-          gherkinQuery.getPickleStepIds(scenarioStepId),
-          pickleStepIds
-        )
+        assert.deepEqual(gherkinQuery.getPickleStepIds(scenarioStepId), pickleStepIds)
       })
     })
   })
@@ -230,10 +213,7 @@ describe('Query', () => {
         }
       },
     })
-    return pipelinePromise(
-      gherkinMessages(gherkinSource, 'test.feature'),
-      writable
-    )
+    return pipelinePromise(gherkinMessages(gherkinSource, 'test.feature'), writable)
   }
 
   function gherkinMessages(gherkinSource: string, uri: string): Readable {