+++ react/javascript/src/components/app/ExecutionSummary.tsx
@@ -1,8 +1,6 @@
 import React from 'react'
 import { messages } from '@cucumber/messages'
-import EnvelopesQueryContext, {
-  EnvelopesQuery,
-} from '../../EnvelopesQueryContext'
+import EnvelopesQueryContext, { EnvelopesQuery } from '../../EnvelopesQueryContext'
 import CICommitLink from './CICommitLink'
 import getDurationsMillis from '../../getDurationMillis'
 import Duration from './Duration'
@@ -12,10 +10,7 @@ interface IProductProps {
   product: messages.Meta.IProduct
 }
 
-const Product: React.FunctionComponent<IProductProps> = ({
-  name: name,
-  product: product,
-}) => {
+const Product: React.FunctionComponent<IProductProps> = ({ name: name, product: product }) => {
   const value = [product.name, product.version]
     .filter((v) => v !== '' && v !== undefined && v !== null)
     .join(' - ')
@@ -28,21 +23,13 @@ const Product: React.FunctionComponent<IProductProps> = ({
   )
 }
 
-function findTestRunStarted(
-  envelopesQuery: EnvelopesQuery
-): messages.ITestRunStarted {
-  const testRunStarted = envelopesQuery.find(
-    (envelope) => envelope.testRunStarted !== null
-  )
+function findTestRunStarted(envelopesQuery: EnvelopesQuery): messages.ITestRunStarted {
+  const testRunStarted = envelopesQuery.find((envelope) => envelope.testRunStarted !== null)
   return testRunStarted ? testRunStarted.testRunStarted : undefined
 }
 
-function findTestRunFinished(
-  envelopesQuery: EnvelopesQuery
-): messages.ITestRunFinished {
-  const testRunFinished = envelopesQuery.find(
-    (envelope) => envelope.testRunFinished !== null
-  )
+function findTestRunFinished(envelopesQuery: EnvelopesQuery): messages.ITestRunFinished {
+  const testRunFinished = envelopesQuery.find((envelope) => envelope.testRunFinished !== null)
   return testRunFinished ? testRunFinished.testRunFinished : undefined
 }
 
@@ -85,9 +72,7 @@ const ExecutionSummary: React.FunctionComponent<IProps> = ({ meta: meta }) => {
               </td>
             </tr>
           )}
-          {meta.implementation && (
-            <Product name="Implementation" product={meta.implementation} />
-          )}
+          {meta.implementation && <Product name="Implementation" product={meta.implementation} />}
           {meta.runtime && <Product name="Runtime" product={meta.runtime} />}
           {meta.os && <Product name="OS" product={meta.os} />}
           {meta.cpu && <Product name="CPU" product={meta.cpu} />}